move range and exec
