+++ src/command_processor.rs
@@ -158,21 +158,19 @@ impl CommandProcessor {
 
     fn get_command_vec() -> Vec<Box<dyn Cmd>> {
         let commands: Vec<Box<dyn Cmd>> = vec![
-            Box::new(crate::cmd::Range::new()),
             Box::new(crate::cmd::Deployments::new()),
             Box::new(crate::cmd::Services::new()),
             Box::new(crate::cmd::ReplicaSets::new()),
             Box::new(crate::cmd::StatefulSets::new()),
             Box::new(crate::cmd::ConfigMaps::new()),
+            Box::new(crate::cmd::Secrets::new()),
+            Box::new(crate::cmd::Jobs::new()),
             Box::new(crate::cmd::Logs::new()),
             Box::new(crate::cmd::Describe::new()),
-            Box::new(crate::cmd::Exec::new()),
             Box::new(crate::cmd::Events::new()),
             Box::new(crate::cmd::Delete::new()),
-            Box::new(crate::cmd::Secrets::new()),
             Box::new(crate::cmd::PortForward::new()),
             Box::new(crate::cmd::PortForwards::new()),
-            Box::new(crate::cmd::Jobs::new()),
             Box::new(crate::command::alias::Alias::new()),
             Box::new(crate::command::alias::Unalias::new()),
             Box::new(crate::command::click::Clear::new()),
@@ -180,14 +178,16 @@ impl CommandProcessor {
             Box::new(crate::command::click::Contexts::new()),
             Box::new(crate::command::click::EnvCmd::new()),
             Box::new(crate::command::click::Quit::new()),
+            Box::new(crate::command::click::Range::new()),
             Box::new(crate::command::click::SetCmd::new()),
             Box::new(crate::command::click::UtcCmd::new()),
+            Box::new(crate::command::exec::Exec::new()),
             Box::new(crate::command::namespaces::Namespace::new()),
             Box::new(crate::command::namespaces::Namespaces::new()),
             Box::new(crate::command::nodes::Nodes::new()),
-            Box::new(crate::command::volumes::PersistentVolumes::new()),
-            Box::new(crate::command::pods::Pods::new()),
             Box::new(crate::command::pods::Containers::new()),
+            Box::new(crate::command::pods::Pods::new()),
+            Box::new(crate::command::volumes::PersistentVolumes::new()),
         ];
         commands
     }
